fix(scoring): preserve time-of-day precision in dated build tiebreaks
Real-world case: "Crash Landed (USA) (Proto) (2009-07-31T111550)" was
beating "Crash Landed (USA) (Proto) (2009-07-31T144710)" — the later,
presumably more complete build of the same day — because
parse_iso_date discarded everything after "T", so both builds packed
to the identical build_date (20090731) and tied on every field of
score_rom's tiebreak tuple. Which one won was then arbitrary rather
than reflecting which was actually the later build.

Widens RomFile.build_date from u32 to u64 and changes its encoding
from YYYYMMDD to YYYYMMDD * 1_000_000 + HHMMSS (000000 when no time
suffix is present), so same-day builds now sort correctly by time.
Adds parse_iso_time() to parse the bare 6-digit "HHMMSS" suffix.

Ripples: score_rom's tuple type (i32, u32, usize) -> (i32, u64, usize)
since the pre-release tier stores build_ord in that slot; its "numbered
proto" sentinel scales up from 99_000_000 to 99_000_000_000_000 to
stay above the new packed range; the three other tiers' rom.revision
now cast to u64 for the same slot. RomFile.build_date gets
#[ts(type = "number | null")] (existing pattern also used for
filesize etc) since ts-rs maps u64 to bigint by default. Also fixes
examples/audit.rs, which hardcoded the old tuple type.

Two regression tests reproducing the exact case, plus updated literal
values in two existing parser tests that asserted the old unscaled
encoding.

+++ src-tauri/examples/audit.rs
@@ -161,7 +161,7 @@ fn scores_summary(
     (pref_name, pref_score, runner_name, runner_score)
 }
 
-fn fmt_score(s: (i32, u32, usize)) -> String {
+fn fmt_score(s: (i32, u64, usize)) -> String {
     format!("{}+{}+{}", s.0, s.1, s.2)
 }
